Alone mode when bot leaves channel
In the current version, when the bot leaves the channel, it gives an error:

2020-02-14T04:57:35+00:00 error on callback func: Uncaught exception: TypeError: Cannot read property 'getClientCount' of undefined at alonemode:31:49 let clients = backend.getCurrentChannel().getClientCount()

With the changes, you do not get the error anymore and if the bot leaves the channel, you also get the benefit of the alone mode. I am a novice user of the bot, so please analyze if there is more complexity needed for the script.

PS.: The embedded file in the latest beta version does not have the discord backend configured.